Refactor: merge pointer diagram event handlers by adding options object
The options object controls each of the individual features like
pointer capture, touchstart.prevent, etc. The code is written so that
the options test is always at the beginning of the line either as

if (options.X) …
if (!options.X) …

That way if I want to show the code, I can "partially evaluate" those
lines and show the resulting simplified code.