Fix: remove indentation from examples
I edit these examples in emacs web-mode, and <body> and <script> both
get indented 2, and <style> gets indented 4. I strip this indentation
while parsing the file so that I can present it without extra
indentation.

It's a bit fragile because it depends on how I have my emacs set up.

Also fix the html escaping and indentation in the constructed iframe.
HTML5 says that <script> and <style> contents should not be
html-escaped.